Move inheritance logic to visitor level. Add example recipes
