+++ src/main/resources/META-INF/rewrite/CompositeDropwizardToSpringBoot.yaml
@@ -1,0 +1,170 @@
+#
+# Copyright 2021 the original author or authors.
+# <p>
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# <p>
+# https://www.apache.org/licenses/LICENSE-2.0
+# <p>
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.CompositeDropwizardToSpringBoot
+displayName: Migration Recipe with Variables
+description: Main recipe that defines variables for child recipes
+recipeList:
+  - org.openrewrite.java.dropwizard.CoreSetup
+  - org.openrewrite.java.dropwizard.MigrateConfigurationClass
+  - org.openrewrite.java.dropwizard.MigrateHealthChecksAndMetrics
+  - org.openrewrite.java.dropwizard.MigrateSecurity
+  - org.openrewrite.java.dropwizard.MigrateHibernate
+  - org.openrewrite.java.dropwizard.MigrateResourcesToSpringJersey
+  - org.openrewrite.java.dropwizard.MigrateTests
+  - org.openrewrite.java.dropwizard.CodeCleanup
+---
+# Core Application Setup
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.CoreSetup
+displayName: Create Spring Boot Application Entry Point
+description: Creates the main Spring Boot application class.
+recipeList:
+  - org.openrewrite.maven.AddDependency:
+      groupId: org.projectlombok
+      artifactId: lombok
+      version: "1.18.30"
+      onlyIfAbsent: true
+  - org.openrewrite.maven.AddManagedDependency:
+      groupId: org.springframework.boot
+      artifactId: spring-boot-dependencies
+      version: "2.7.18"
+      scope: import
+      type: pom
+      addToRootPom: true
+  - org.openrewrite.general.annotation.AddAnnotationIfSuperTypeExistsRecipe:
+      annotationToAdd: "org.springframework.boot.SpringBootApplication"
+      targetSupertypeNames: "io.dropwizard.Application"
+
+---
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.MigrateHealthChecksAndMetrics
+displayName: Add Spring Boot Actuator
+description: Configures Spring Boot Actuator with basic health endpoints.
+recipeList:
+  - org.openrewrite.maven.AddDependency:
+      groupId: org.springframework.boot
+      artifactId: spring-boot-starter-actuator
+      version: "2.7.18"
+      onlyIfAbsent: true
+
+---
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.ModifyDropwizardHealthChecksToSpringVariants
+displayName: Convert Health Check Implementations
+description: Transforms Dropwizard HealthCheck classes to Spring Boot HealthIndicator.
+recipeList:
+  - org.openrewrite.general.annotation.AddAnnotationIfSuperTypeExistsRecipe:
+      annotationToAdd: "org.springframework.stereotype.Component"
+      targetSupertypeNames: "org.springframework.boot.actuate.health.HealthIndicator"
+---
+# Configuration Migration
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.MigrateConfigurationClass
+displayName: Migrate Configuration Class
+description: Converts Dropwizard Configuration to Spring Boot format.
+recipeList:
+  - org.openrewrite.general.annotation.AddAnnotationIfSuperTypeExistsRecipe:
+      annotationToAdd: "org.springframework.context.annotation.Configuration"
+      targetSupertypeNames: "io.dropwizard.Configuration"
+  - org.openrewrite.general.annotation.AddAnnotationIfSuperTypeExistsRecipe:
+      annotationToAdd: "org.springframework.boot.context.properties.ConfigurationProperties"
+      targetSupertypeNames: "io.dropwizard.Configuration"
+  - org.openrewrite.general.annotation.AddAnnotationIfSuperTypeExistsRecipe:
+      annotationToAdd: "lombok.Data"
+      targetSupertypeNames: "io.dropwizard.Configuration"
+      annotateInnerClasses: true
+  - org.openrewrite.general.annotation.AddAnnotationIfAnnotationExistsRecipe:
+      annotationToAdd: "org.springframework.stereotype.Component"
+      targetAnnotationClassNames: "javax.ws.rs.ext.Provider"
+---
+# Resource Migration
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.MigrateResourcesToSpringJersey
+displayName: Migrate Resource Classes
+description: Converts Dropwizard Resources to Spring Boot format.
+recipeList:
+  - org.openrewrite.maven.AddDependency:
+      groupId: org.springframework.boot
+      artifactId: spring-boot-starter
+      version: "2.7.18"
+      onlyIfAbsent: true
+  - org.openrewrite.maven.AddDependency:
+      groupId: org.springframework.boot
+      artifactId: spring-boot-starter-web
+      version: "2.7.18"
+      onlyIfAbsent: true
+  - org.openrewrite.maven.AddDependency:
+      groupId: org.springframework.boot
+      artifactId: spring-boot-starter-jersey
+      version: "2.7.18"
+      onlyIfAbsent: true
+  - org.openrewrite.general.annotation.AddAnnotationIfAnnotationExistsRecipe:
+      annotationToAdd: "org.springframework.stereotype.Component"
+      targetAnnotationClassNames: "javax.ws.rs.Path"
+---
+# Hibernate Migration
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.MigrateHibernate
+displayName: Migrate Hibernate
+description: Converts Dropwizard Resources to Spring Boot format.
+recipeList:
+  - org.openrewrite.maven.AddDependency:
+      groupId: org.springframework.boot
+      artifactId: spring-boot-starter-data-jpa
+      version: "2.7.18"
+      onlyIfAbsent: true
+  - org.openrewrite.general.annotation.AddAnnotationIfSuperTypeExistsRecipe:
+      annotationToAdd: "org.springframework.stereotype.Repository"
+      targetSupertypeNames: "io.dropwizard.hibernate.AbstractDAO"
+  - org.openrewrite.general.annotation.AddAnnotationIfSuperTypeExistsRecipe:
+      annotationToAdd: "org.springframework.transaction.annotation.Transactional"
+      targetSupertypeNames: "io.dropwizard.hibernate.AbstractDAO"
+---
+# Test migration
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.MigrateTests
+displayName: Migrate Health Checks to Spring Boot
+description: Converts Dropwizard tests to Spring Boot format.
+recipeList:
+  - org.openrewrite.maven.AddDependency:
+      groupId: org.springframework.boot
+      artifactId: spring-boot-starter-test
+      version: "2.7.18"
+      onlyIfAbsent: true
+      scope: test
+---
+# Security migration
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.MigrateSecurity
+displayName: Migrate Health Checks to Spring Boot
+description: Converts Dropwizard health checks to Spring Boot format.
+recipeList:
+  - org.openrewrite.maven.AddDependency:
+      groupId: org.springframework.boot
+      artifactId: spring-boot-starter-security
+      version: "2.7.18"
+      onlyIfAbsent: true
+---
+# Cleanup
+type: specs.openrewrite.org/v1beta/recipe
+name: org.openrewrite.java.dropwizard.CodeCleanup
+displayName: Clean up various issues with the code.
+description: Clean up the code.
+recipeList:
+  - org.openrewrite.java.ShortenFullyQualifiedTypeReferences
+  - org.openrewrite.general.method.AddMissingAbstractMethods
+  - org.openrewrite.java.RemoveUnusedImports
